Decouple the Twig extension from the Templating component
