+++ Resources/config/renderer.xml
@@ -5,9 +5,8 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
     <services>
-        <service id="ivory_ck_editor.templating.helper" class="Ivory\CKEditorBundle\Templating\CKEditorHelper">
-            <argument type="service" id="ivory_ck_editor.renderer" />
-            <tag name="templating.helper" alias="ivory_ckeditor" />
+        <service id="ivory_ck_editor.renderer" class="Ivory\CKEditorBundle\Renderer\CKEditorRenderer">
+            <argument type="service" id="service_container" />
         </service>
     </services>
 </container>
