+++ Tests/IvoryCKEditorBundleTest.php
@@ -21,33 +21,62 @@ use Ivory\CKEditorBundle\IvoryCKEditorBundle;
  */
 class IvoryCKEditorBundleTest extends \PHPUnit_Framework_TestCase
 {
+    /** @var \Ivory\CKEditorBundle\IvoryCKEditorBundle */
+    private $bundle;
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp()
+    {
+        $this->bundle = new IvoryCKEditorBundle();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function tearDown()
+    {
+        unset($this->bundle);
+    }
+
     public function testBundle()
     {
-        $bundle = new IvoryCKEditorBundle();
-
-        $this->assertInstanceOf('Symfony\Component\HttpKernel\Bundle\Bundle', $bundle);
+        $this->assertInstanceOf('Symfony\Component\HttpKernel\Bundle\Bundle', $this->bundle);
     }
 
     public function testCompilerPasses()
     {
-        $containerBuilder = $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerBuilder')
-            ->disableOriginalConstructor()
-            ->setMethods(array('addCompilerPass'))
-            ->getMock();
-
+        $containerBuilder = $this->createContainerBuilderMock();
         $containerBuilder
             ->expects($this->at(0))
             ->method('addCompilerPass')
-            ->with($this->isInstanceOf('Ivory\CKEditorBundle\DependencyInjection\Compiler\ResourceCompilerPass'))
+            ->with($this->isInstanceOf('Ivory\CKEditorBundle\DependencyInjection\Compiler\AssetsHelperCompilerPass'))
             ->will($this->returnSelf());
 
         $containerBuilder
             ->expects($this->at(1))
             ->method('addCompilerPass')
-            ->with($this->isInstanceOf('Ivory\CKEditorBundle\DependencyInjection\Compiler\AssetsHelperCompilerPass'))
+            ->with($this->isInstanceOf('Ivory\CKEditorBundle\DependencyInjection\Compiler\ResourceCompilerPass'))
             ->will($this->returnSelf());
 
-        $bundle = new IvoryCKEditorBundle();
-        $bundle->build($containerBuilder);
+        $containerBuilder
+            ->expects($this->at(2))
+            ->method('addCompilerPass')
+            ->with($this->isInstanceOf('Ivory\CKEditorBundle\DependencyInjection\Compiler\TemplatingCompilerPass'))
+            ->will($this->returnSelf());
+
+        $this->bundle->build($containerBuilder);
+    }
+
+    /**
+     * @return \Symfony\Component\DependencyInjection\ContainerBuilder|\PHPUnit_Framework_MockObject_MockObject
+     */
+    private function createContainerBuilderMock()
+    {
+        return $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerBuilder')
+            ->disableOriginalConstructor()
+            ->setMethods(array('addCompilerPass'))
+            ->getMock();
     }
 }
