+++ Tests/Template/TwigTemplateTest.php
@@ -34,7 +34,7 @@ class TwigTemplateTest extends AbstractTemplateTest
         parent::setUp();
 
         $this->twig = new \Twig_Environment(new \Twig_Loader_Filesystem(array(__DIR__.'/../../Resources/views/Form')));
-        $this->twig->addExtension(new CKEditorExtension($this->helper));
+        $this->twig->addExtension(new CKEditorExtension($this->renderer));
 
         $this->template = $this->twig->loadTemplate('ckeditor_widget.html.twig');
     }
